% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetLibR-package.R
\docType{package}
\name{NetLibR}
\alias{NetLibR}
\alias{NetLibR-package}
\title{NetLibR: An R interface for a set of useful graph algorithms}
\description{
Graph algorithms for clustering, motif detection, alignment, ...
}
\section{Usage}{
 

\enumerate{
\item ?signed.cluster
}
}

\section{Useful links}{


\enumerate{
\item Report bugs at \url{https://github.com/shmohammadi86/GraphLibR/issues}
\item Read the manuscript: 
\href{https://}{Soon!}.
}
}

